pyproject.toml
[tool.poetry]
name = "dune-lt"
version = "0.1.0"
description = ""
authors = ["Steven Wang <steven@definite.app>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
dlt = "^1.6.1"
notebook = "^7.3.2"
duckdb = "^1.2.0"
dune-spice = "^0.2.6"
pandas = "^2.2.3"
streamlit = "^1.42.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
